requirements.txt
asgiref==3.9.1
Django==5.2.4
django-grappelli==4.0.2
djangorestframework==3.16.0
python-decouple==3.8
sqlparse==0.5.3
